requirements.txt
aiogram==3.*
httpx==0.27.*
python-dotenv==1.0.1